pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "machine-rules"
version = "0.1.0"
description = "JSR-94 compatible rules engine for Python"
authors = [
    { name="Your Name", email="your.email@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "pyyaml",
    "durable-rules",
    "pydantic",
    "pytest",
    "pandas",
    "openpyxl"
]

[tool.setuptools.packages.find]
where = ["."]
